+++ Graphic11/base/dxInit.h
@@ -6,8 +6,8 @@ namespace graphic
 	// Direct3D11 Initialize
 	bool InitDirectX11(
 		const HWND hWnd
-		, const int width
-		, const int height
+		, const float width
+		, const float height
 		, OUT ID3D11Device **pd3dDevice
 		, OUT ID3D11DeviceContext **pImmediateContext
 		, OUT IDXGISwapChain **pSwapChain
@@ -19,8 +19,8 @@ namespace graphic
 	bool InitDirectX11SwapChain(
 		ID3D11Device *pd3dDevice
 		, const HWND hWnd
-		, const int width
-		, const int height
+		, const float width
+		, const float height
 		, OUT IDXGISwapChain **pSwapChain
 		, OUT ID3D11RenderTargetView **pRenderTargetView
 		, OUT ID3D11DepthStencilView **pDepthStencilView
@@ -31,8 +31,8 @@ namespace graphic
 	bool InitDirectX11Deferred(
 		ID3D11Device *pd3dDevice
 		, const HWND hWnd
-		, const int width
-		, const int height		
+		, const float width
+		, const float height
 		, OUT ID3D11DeviceContext **pDeferredContext
 		, OUT IDXGISwapChain **pSwapChain
 		, OUT ID3D11RenderTargetView **pRenderTargetView
